chore(envrc): add .envrc file to use asdf version manager
docs(README.md): update caution and support sections for clarity

Add a new .envrc file to utilize the asdf version manager, which helps in managing different versions of programming languages and tools. This change is aimed at improving the development environment setup.

In the README.md, update the caution section to clarify the unofficial status of the image and emphasize the lack of support from CubeCoders and the maintainer. Remove redundant sections and streamline the document to focus on essential information, making it easier for users to understand the risks and responsibilities associated with using the image.