docs(README.md): update cautionary notes for clarity and emphasis
Change the cautionary note from "CAUTION" to "WARNING" to better
convey the seriousness of the message. Add a separate "CAUTION" tag
for the "USE AT YOUR OWN RISK" statement to emphasize the potential
risks involved. These changes aim to ensure users are fully aware
of the unofficial nature of the image and the lack of support.