chore(deploy-prod.yml): remove debug step and commented-out login step
Remove the debug step that outputs the GITHUB_ACTOR as it is no longer
necessary for the deployment process. Also, remove the commented-out
login step to GitHub Container Registry to clean up the workflow file
and avoid confusion. This streamlines the deployment workflow, making
it more efficient and easier to maintain.